[module4/class-2] - createing fetch product in product detail
